zsh,alacritty: resolved alacritty/tmux/pywal bug
When tmux was launched the background wasn't transparent.

The problem was in zshrc where I use the command given on the wiki.
I don't know why this happenned but commenting out the line resolve the problem.

However, I lost pywal colour palette. Searching the issues on the github I found
a user that propose some workaround for alacritty and I use that.
The source is on the first line of the alacritty.yml file.