script: new session_log use only script
This is a tmp workaround.
As script and zellij launch their own subshell.
I can seems to be able to launch script and then zellij through a script.
That's why I decided to first make something workable with script then I will
make something for zellij.
Hopefully, after that I may be able to combine the two.
Wait and see..